Disabled a parent lookup in setEnabled call
This lookup was introduced during a bug fix. It will break a "visible" Comparison where the subject is a checkbox that is being enhanced using the bootstrapSwitch plugin.
Disabling the line doesn't seem to have a negative affect on the revision it was introduced in. Bug #26 is still fixed. Still needs more testing.